--- specification/ml/_types/DataCounts.ts
@@ -1,42 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { Id } from '@_types/common'
-import { long } from '@_types/Numeric'
-
-export class DataCounts {
-  bucket_count: long
-  earliest_record_timestamp?: long
-  empty_bucket_count: long
-  input_bytes: long
-  input_field_count: long
-  input_record_count: long
-  invalid_date_count: long
-  job_id: Id
-  last_data_time?: long
-  latest_empty_bucket_timestamp?: long
-  latest_record_timestamp?: long
-  latest_sparse_bucket_timestamp?: long
-  latest_bucket_timestamp?: long
-  missing_field_count: long
-  out_of_order_timestamp_count: long
-  processed_field_count: long
-  processed_record_count: long
-  sparse_bucket_count: long
-}
