+++ specification/ml/_types/Functions.ts
@@ -17,8 +17,62 @@
  * under the License.
  */
 
+export enum CountFunction {
+  Count = 0,
+  HighCount = 1,
+  LowCount = 2
+}
+
+export enum DistinctCountFunction {
+  DistinctCount = 0,
+  LowDistinctCount = 1,
+  HighDistinctCount = 2
+}
+
+export enum GeographicFunction {
+  LatLong = 0
+}
+
+export enum InfoContentFunction {
+  InfoContent = 0,
+  HighInfoContent = 1,
+  LowInfoContent = 2
+}
+
+export enum MetricFunction {
+  Min = 0,
+  Max = 1,
+  Median = 2,
+  HighMedian = 3,
+  LowMedian = 4,
+  Mean = 5,
+  HighMean = 6,
+  LowMean = 7,
+  Metric = 8,
+  Varp = 9,
+  HighVarp = 10,
+  LowVarp = 11
+}
+
 export enum NonNullSumFunction {
   NonNullSum = 0,
   HighNonNullSum = 1,
   LowNonNullSum = 2
+}
+
+export enum NonZeroCountFunction {
+  NonZeroCount = 0,
+  LowNonZeroCount = 1,
+  HighNonZeroCount = 2
+}
+
+export enum RareFunction {
+  Rare = 0,
+  FreqRare = 1
+}
+
+export enum SumFunction {
+  Sum = 0,
+  HighSum = 1,
+  LowSum = 2
 }
