+++ specification/ml/upgrade_job_snapshot/MlUpgradeJobSnapshotResponse.ts
@@ -17,10 +17,12 @@
  * under the License.
  */
 
+import { NodeId } from '@_types/common'
+
 export class Response {
   body: {
     /** The ID of the assigned node for the upgrade task if it is still running. */
-    node: string
+    node: NodeId
     /** When true, this means the task is complete. When false, it is still running. */
     completed: boolean
   }
